Parsing build ok

+++ srcs/init/init_data.c
@@ -6,7 +6,7 @@
 /*   By: mfeldman <mfeldman@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/24 14:49:44 by fllanet           #+#    #+#             */
-/*   Updated: 2024/02/29 12:22:10 by mfeldman         ###   ########.fr       */
+/*   Updated: 2024/03/01 16:45:33 by mfeldman         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 void	clean_nl_scene(t_data *data)
 {
-	int		i; // uint, if f2_scene ok, not more than 6 lines 
-	int		j;
+	uint8_t		i;
+	int			j;
 
 	i = 0;
 	while (data->infos[i])
@@ -72,10 +72,6 @@ bool	init_images(t_data *data)
 
 bool	init_data(t_data *data)
 {
-	// t_image	image;
-	
-	// image = (t_image){0};
-	// data->img = &image;
 	get_player_pos(data); //recode with fct in parsing 
 	data->mlx = mlx_init();
 	if (!data->mlx)
@@ -84,24 +80,3 @@ bool	init_data(t_data *data)
 		return (data->error->error_g |= ERROR_INIT_IMG, 1);
 	return (0);
 }
-
-// bool	init_struct(t_data *data)
-// {
-// 	// data->error = (t_error *)malloc(sizeof(t_error));
-// 	// if (!data->error)
-// 	// 	return (1);
-// 	// ft_memset(data->error, 0, sizeof(t_error));
-// 	data->game = (t_game *)malloc(sizeof(t_game));
-// 	if (!data->game)
-// 		return (1);
-// 	ft_memset(data->game, 0, sizeof(t_game));
-// 	data->ray = (t_raycast *)malloc(sizeof(t_raycast));
-// 	if (!data->ray)
-// 		return (1);
-// 	ft_memset(data->ray, 0, sizeof(t_raycast));
-// 	data->img = (t_image *)malloc(sizeof(t_image));
-// 	if (!data->img)
-// 		return (1);
-// 	ft_memset(data->img, 0, sizeof(t_image));
-// 	return (0);
-// }
